refactor: Extract shared manifest-schema module
Create src/manifest-schema/ as shared infrastructure consumed by both
manifest-intellisense and manifest-editor:

- schema-model.ts: Type definitions and namespace constants (moved)
- xsd-parser.ts: XSD parser (moved)
- xml-context.ts: XML cursor context analysis (moved)
- schema-validation.ts: Schema-driven validation (extracted from intellisense-logic.ts)
- index.ts: Public API barrel export

Refactor manifest-intellisense to import from shared module:
- Replace local schema-model.ts, xsd-parser.ts, xml-context.ts with re-exports
- Remove duplicated validation code from intellisense-logic.ts
- Re-export validateManifestText/findSchemaElementExact for backward compat

Refactor manifest-editor to use XSD schema for validation:
- Add schema-helpers.ts with validateValueAgainstType(), matchesSchemaPattern(),
  isValidSchemaColor(), buildAttributeFromPatternType()
- Update validateManifest() to accept optional SchemaModel parameter
- Replace hand-written regexes with schema lookups when schema available:
  IDENTITY_NAME_REGEX -> ST_PackageName, VERSION_REGEX -> ST_VersionQuad,
  PUBLISHER_DN_REGEX -> ST_Publisher_2010_v2, GUID_REGEX -> ST_GUID,
  HEX_COLOR_REGEX+NAMED_COLORS -> ST_Color, APP_ID_REGEX -> ST_ApplicationId
- Keep hand-written regexes as fallback when no schema loaded
- Semantic rules stay as-is (reserved names, version comparison, image checks)

Consolidate namespace constants:
- xml-utils.ts NS map now imports from shared MANIFEST_NAMESPACES
- Extension-specific namespaces (win32dependencies, systemai) kept as literals

Wire shared schema in extension.ts:
- Single lazy-loaded SchemaModel shared between editor and IntelliSense
- Both modules receive the same getSharedSchema() getter

Add 26 new tests in schema-shared.test.ts covering:
- Barrel exports, schema-validation functions, schema-helpers integration
- Pattern type inheritance resolution, color validation, length checks
- validateManifest with and without schema parameter

All 436 relevant tests passing.

Copilot-Session: 37e9b215-2ad4-44e8-abc7-3073d1c4edc0
