+++ src/manifest-editor/xml-utils.ts
@@ -2,16 +2,20 @@
  * Low-level XML string manipulation utilities for the manifest editor.
  */
 
-// Common AppxManifest namespace URIs
+import { MANIFEST_NAMESPACES } from '../manifest-schema/schema-model';
+
+// Namespace constants derived from the shared schema model.
+// NS provides backward-compatible named access for the manifest editor.
 export const NS = {
-    default: 'http://schemas.microsoft.com/appx/manifest/foundation/windows10',
-    uap: 'http://schemas.microsoft.com/appx/manifest/uap/windows10',
-    uap3: 'http://schemas.microsoft.com/appx/manifest/uap/windows10/3',
-    uap5: 'http://schemas.microsoft.com/appx/manifest/uap/windows10/5',
-    uap7: 'http://schemas.microsoft.com/appx/manifest/uap/windows10/7',
-    uap10: 'http://schemas.microsoft.com/appx/manifest/uap/windows10/10',
-    rescap: 'http://schemas.microsoft.com/appx/manifest/foundation/windows10/restrictedcapabilities',
-    desktop: 'http://schemas.microsoft.com/appx/manifest/desktop/windows10',
+    default: MANIFEST_NAMESPACES[''],
+    uap: MANIFEST_NAMESPACES['uap'],
+    uap3: MANIFEST_NAMESPACES['uap3'],
+    uap5: MANIFEST_NAMESPACES['uap5'],
+    uap7: MANIFEST_NAMESPACES['uap7'],
+    uap10: MANIFEST_NAMESPACES['uap10'],
+    rescap: MANIFEST_NAMESPACES['rescap'],
+    desktop: MANIFEST_NAMESPACES['desktop'],
+    // These don't exist in the IntelliSense schema set — keep as literals
     win32dependencies: 'http://schemas.microsoft.com/appx/manifest/win32dependencies/windows10',
     systemai: 'http://schemas.microsoft.com/appx/manifest/systemai/windows10',
 };
